Initial commit - action and script to tag/upload/replace ref
An action which takes a number of inputs to specify which bucket,
role and ECR repo to push to as well as a KMS key, and the
working directory of the containers to be pushed.
Also a script which does the actual building and tagging then
replaces the references in the template file and uploads it to
the artifact bucket(s).